+++ src/core/function/ObjectDefine.ts
@@ -3,11 +3,11 @@
  * @author: Wibus
  * @Date: 2021-09-25 16:45:46
  * @LastEditors: Wibus
- * @LastEditTime: 2021-09-25 16:48:43
+ * @LastEditTime: 2021-09-25 21:50:29
  * Coding With IU
  */
-function objAdd(data: any, where: PropertyKey, value: any, writable = false){
-    Object.defineProperty(data, where, {
+function objAdd(data: any, name: PropertyKey, value: any, writable = false){
+    Object.defineProperty(data, name, {
         value: value,
         writable: writable,
     })