+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Features
 
 * add controller ([843735f](https://github.com/wibus-wee/Nest-server/commit/843735f2f93e0ec63c3792c1af803f1d1a46164f))
+* **controller:** improve the Host and Friends controllers ([04b7f9f](https://github.com/wibus-wee/Nest-server/commit/04b7f9f42a573dbe037f390f0f3beb7727dfe695))
 * project init ([25ba497](https://github.com/wibus-wee/Nest-server/commit/25ba497fb2d3b31cffa68a762fec1433fab47f60))